major refactoring
removed most of the "helper" methods from Map's interface to simplify it,
and re-implemented the internals that relied on them

map state tracking is now handled by an internal object of map

better handling of magic strings for file and directory names

resolves #85
resolves #93
resolves #103
